requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.7.1
pycryptodome==3.21.0
python-dotenv==1.0.1
werkzeug==3.1.3
pillow==11.1.0
